# src/swagger.json
{
	"basePath": "/api",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Name": {
			"properties": {
				"first": {
					"type": "string"
				},
				"last": {
					"type": "string"
				}
			},
			"required": [
				"first"
			],
			"type": "object"
		},
		"User": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double"
				},
				"email": {
					"type": "string"
				},
				"name": {
					"$ref": "#/definitions/Name"
				},
				"status": {
					"type": "string",
					"enum": [
						"Happy",
						"Sad"
					]
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"id",
				"email",
				"name",
				"phoneNumbers"
			],
			"type": "object"
		},
		"UserCreationRequest": {
			"properties": {
				"email": {
					"type": "string"
				},
				"name": {
					"$ref": "#/definitions/Name"
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"email",
				"name",
				"phoneNumbers"
			],
			"type": "object"
		},
		"Test": {
			"properties": {
				"id": {
					"type": "number",
					"format": "double"
				},
				"name": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name"
			],
			"type": "object"
		},
		"TestCreationRequest": {
			"properties": {
				"name": {
					"type": "string"
				}
			},
			"required": [
				"name"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "mern-ts-api-server-template",
		"version": "1.0.6",
		"description": "Template for a full-featured MERN stack/Typescript & Swagger/InversifyJS APIs server with nodemon to monitor for any changes in your source and automatically restart Node server",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/v1/users/{id}": {
			"get": {
				"operationId": "UsersGetUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					},
					{
						"in": "query",
						"name": "name",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/v1/users": {
			"post": {
				"operationId": "UsersCreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserCreationRequest"
						}
					}
				]
			}
		},
		"/v1/test/{id}": {
			"get": {
				"operationId": "TestGetTest",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/Test"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					},
					{
						"in": "query",
						"name": "name",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/v1/test": {
			"post": {
				"operationId": "TestCreateTest",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/TestCreationRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {},
	"host": "localhost:3000"
}